Use secret manager from latest deployment snapshot in `preview`s and `up`s
If `preview`/`up` is running on a stack that does not have a
`Pulumi.stack-name.yaml` with a configured secret provider, it should
use the secret provider available in the last deployment snapshot.

This is useful in cases where `Pulumi.stack-name.yaml` are not committed
to source control or when the provider is not specified in them.

Before this commit it will default to the Pulumi Cloud managed secret
provider.

`getStackSecretsManager()` was fixed to now return the secret manager
from the state if it exists.

In addition:
* Now `pulumi config refresh` will not fail if there are
  no previous deployments.
* `pulumi stack init --secrets-provider` will now save the secrets
  provider immediately to the state.